From d3e0bf403d527f717a62ea328781256f999d4f95 Mon Sep 17 00:00:00 2001 Subject: [PATCH] mlxsw: i2c: Prevent transaction execution for special chip states
Do not run transaction in cases chip is in reset or in-service update
states. In such case firmware is not accessible and will reject transaction
with the relevant status "RUNNING_RESET" or "FW_ISSU_ONGOING".
In case transaction is failed do to one of these reasons, stop sending
transactions. In such case driver is about to be removed since it
cannot continue running after reset or in-service update. And
re-probed again after reset or in-service update is completed.

Signed-off-by: Vadim Pasternak <vadimp@nvidia.com>
Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>