mlxsw: i2c: Fix chunk size setting in output mailbox buffer
Set output mailbox buffer size multiple of minimal chunk size (32).

Full buffer size is 256 bytes, while chunk size, which can be sent to
device on some controllers could be for example 32 + 4, 64 + 4, 128 +
4. Thus, last chunk maybe missed, and transaction tail will be lost.

For example, if transaction size is 256 bytes and chunk size is 64 + 4,
only 204 (68 * 3) bytes will be read instead of 256.

With this fix chunk size will be multiple of 2^n (where n could be 5, 6,
7) and last chunk will be handled.

Fixes: 95b75cbd1bc5 ("mlxsw: i2c: Extend input parameters list of command API")
Signed-off-by: Vadim Pasternak <vadimp@nvidia.com>